+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 0.1.5 (unreleased)
 
-* TBD
+IMPROVEMENTS
+
+* Add configuration option `metrics.graphite.min-connect-attempt-interval-secs`, which configures the minimum wait time
+  (in seconds) in between connection attempts to Graphite.
 
 
 # 0.1.4 (March 04, 2015)